==> 02-notebooks/03-collaborative-filter/01-criacao-dados-als.ipynb <==
# Criação dos dados para o ALS

## 0. Setup

#---- Bibliotecas:

# Análise e manipulação de dados:

import pandas as pd
import numpy as np

# Visualização de dados:

import plotly.io as pio
import plotly.express as px

# Manuseio nos dados

from os import chdir

#---- Mudando o diretório raiz para a pasta principal do projeto

chdir('../../')

#---- Template dos gráficos:

pio.templates.default = "plotly_white"

#---- Carregando as funções

from functions import *

## 1. Extração dos dados

#---- Dados de treino:

df_train = pd.read_parquet('03-data/02-processed/01-train_data.parquet')

df_train.head()

## 2. Manuseio nos dados

### 2.1. Criação das variáveis de ID para entrar na matriz usuário X item

df_train['USER_ID_CAT'] = df_train['User-ID'].astype('category').cat.codes
df_train['IBSN_ID_CAT'] = df_train['ISBN'].astype("category").cat.codes

df_dim_user = df_train[['User-ID', 'USER_ID_CAT']]\
    .drop_duplicates()\
    .reset_index(drop = True)\
    .reset_index()\
    .assign(\
        USER_ID_INT = lambda x: x['index'] + 1
    )\
    .drop(columns = 'index')

df_dim_user

df_dim_isbn = df_train[['ISBN', 'IBSN_ID_CAT']]\
    .drop_duplicates()\
    .reset_index(drop = True)\
    .reset_index()\
    .assign(\
        ISBN_ID_INT = lambda x: x['index'] + 1
    )\
    .drop(columns = 'index')

df_dim_isbn.head()

df_train = df_train\
    .merge(df_dim_user, on = ['User-ID', 'USER_ID_CAT'], how = 'left')\
    .merge(df_dim_isbn, on = ['ISBN', 'IBSN_ID_CAT'], how = 'left')

df_train.head()

#---- Criando a matriz de usuário x item

sparse_matrix = csr_matrix(
    (
        df_train['Book-Rating'], 
        (df_train['USER_ID_INT'], df_train['ISBN_ID_INT'])
    )
)

sparse_matrix

## 3. Salvando os dados

#---- Matriz usuário e item

save_sparse_csr(array = sparse_matrix, filename = '03-data/02-processed/04-ALS-train-data.npz')

#---- Dados de treino

df_train\
    .to_parquet('03-data/02-processed/05-ALS-df-train-data.parquet', index = False)

# #---- DE-PARA

# df_dim_isbn\
#     .to_parquet('03-data/02-processed/05-ALS-DE-PARA-ISBN.parquet', index = False)

# df_dim_user\
#     .to_parquet('03-data/02-processed/06-ALS-DE-PARA-USER.parquet', index = False)
