==> 02-notebooks/03-collaborative-filter/02-als.ipynb <==
## ALS

## 0. Setup

%load_ext autotime

#---- Bibliotecas:

# Análise e manipulação de dados:

import pandas as pd
import swifter
import numpy as np

# Visualização de dados:

import plotly.io as pio
import plotly.express as px

# Manuseio nos dados

from os import chdir

# ALS

from implicit import als

#---- Mudando o diretório raiz para a pasta principal do projeto

chdir('../../')

#---- Template dos gráficos:

pio.templates.default = "plotly_white"

#---- Carregando as funções

from functions import *

## 1. Extração dos dados

#---- Dados de treino

train_data = load_sparse_csr('03-data/02-processed/04-ALS-train-data.npz')

train_data

#---- Dados de teste

df_test = pd.read_parquet('03-data/02-processed/02-test_data.parquet')

df_test.head()

df_test.info()

#---- Dados de treino

df_train = pd.read_parquet('03-data/02-processed/05-ALS-df-train-data.parquet')

df_train.head()

# #---- De-para's:

# de_para_isbn = pd.read_parquet('03-data/02-processed/05-ALS-DE-PARA-ISBN.parquet')
# de_para_user = pd.read_parquet('03-data/02-processed/06-ALS-DE-PARA-USER.parquet')

# #---- Full data:

# df_full = pd.read_parquet('03-data/02-processed/03-full_data.parquet')

# df_full.head()

## 2. Modelagem

model = als.AlternatingLeastSquares(factors = 30, random_state = 42)

model

model.fit(train_data)

#---- DE-PARA USER_ID e ISBN:

de_para_user_id = dict(zip(df_train['User-ID'].tolist(), df_train['USER_ID_INT'].tolist()))
de_para_isbn = dict(zip(df_train['ISBN'].tolist(), df_train['ISBN_ID_INT'].tolist()))

## 3. Recomendações

user_id_rec = 276747

de_para_user_id_rec = de_para_user_id[user_id_rec]

de_para_user_id_rec

# #---- TOP 10 livros mais bem avaliados pelo usuário

# df_train\
#     .query(f'`User-ID` == {user_id_rec}')\
#     .query(f'`Book-Rating` != 0')\
#     .sort_values(by = 'Book-Rating', ascending = False)[['ISBN', 'Book-Title', 'Book-Rating']]\
#     .drop_duplicates()\
#     .head(10)

# #---- Lista de ISBN que o usuário já leu

# readed_isbn_by_user = df_train\
#     .query(f'`User-ID` == {user_id_rec}')['Book-Title']\
#     .unique()\
#     .tolist()

# readed_isbn_by_user

# not_recommended_list_to_user = df_train\
#     .query(f'`Book-Title`.isin({readed_isbn_by_user})')['ISBN_ID_INT']\
#     .unique()\
#     .tolist()

# #---- Top10 recomendações para o usuário

# recommendations = model.recommend(de_para_user_id_rec,
#                                   train_data[de_para_user_id_rec], 
#                                   filter_already_liked_items = True,
#                                   N = 10,
#                                   filter_items = not_recommended_list_to_user)

# print(f'Recomendações: {recommendations}')

# list_recommendations = recommendations[0].tolist()

# #---- De-para das recomendações para ISBN

# # isbn_recommendations = df_train\
# #     .query(f'ISBN_ID_INT.isin({list_recommendations})')['ISBN']\
# #     .unique()\
# #     .tolist()

# isbn_recommendations = list({k: v for k, v in de_para_isbn.items() if v in list_recommendations}.keys())

# df_train\
#     .pipe(filter_list_ordered, 
#           col = 'ISBN', 
#           list_filter = isbn_recommendations)\
#     [['ISBN', 'Book-Title']]\
#     .drop_duplicates()['Book-Title'].unique()

## 4. Avaliação das métricas

# Criar uma função que olhe a coluna User-ID e faça as TOP N recomendações. Resultado: lista
# Validar se dentro dessa lista de recomendações, está o livro-teste
# Criaremos uma flag: se o livro está dentro das recomendações, 1, senão 0
# Proporção dos valores 1 (Precision @ K)

#---- Criando uma função para recomendar K itens a partir do nome do usuário

def choose_books_als(user_id_to_rec: str, dict_de_para_user: dict, dict_de_para_isbn: dict, rec_matrix, n: int = 10):

    # User_id para fazer a recomendação:

    user_id_de_para = dict_de_para_user[user_id_to_rec]

    recommendations = model.recommend(user_id_de_para,
                                      rec_matrix[user_id_de_para], 
                                      filter_already_liked_items = True,
                                      N = n)

    list_recommendations = recommendations[0].tolist()

    #---- De-para das recomendações para ISBN

    isbn_recommendations = list({k: v for k, v in de_para_isbn.items() if v in list_recommendations}.keys())

    return isbn_recommendations

choose_books_als(user_id_to_rec = 276747,
                 dict_de_para_user = de_para_user_id,
                 dict_de_para_isbn = de_para_isbn,
                 rec_matrix = train_data)

# df_test_copy = df_test\
#     .sample(1000)

df_test['TOP_N_RECOMMENDATIONS'] = df_test\
    .swifter.apply(lambda x: choose_books_als(user_id_to_rec = x['User-ID'], 
                                      dict_de_para_user = de_para_user_id,
                                      dict_de_para_isbn = de_para_isbn,
                                      rec_matrix = train_data), axis = 1)

df_test\
    .head(5)

## 5. Validando o Precision @  K

def calculate_precision_k_als(df: pd.DataFrame, col_books: str):

    df['flag_recommendations'] = df.apply(lambda x: x[col_books] in x['TOP_N_RECOMMENDATIONS'], axis = 1)

    value_precision_k = df.query('flag_recommendations == 1').shape[0] / df.shape[0] * 100

    precision_k = 0 if value_precision_k == 0 else value_precision_k

    return df, precision_k

calculate_precision_k_als(df = df_test, col_books = 'Book-Title')[0]

# Para a próxima rodada de estudo:

- Otimizar o código
- Continuar a avaliação da métrica

# TO-DO's

- Ler um pouco mais sobre a teoria do ALS
- Criar um código que exclua o primeiro item da recomendação do usuário

# Validações

df_train['FLAG_HP'] = np.where(df_train['Book-Title'].str.contains('Potter'), 1, 0)

test = df_train[['User-ID', 'FLAG_HP']]\
    .value_counts()\
    .reset_index()\
    .sort_values(by = ['User-ID', 'FLAG_HP'])

test['percent'] = test['count'] / test.groupby('User-ID')['count'].transform('sum') * 100

test\
    .sort_values(by = 'percent', ascending = False)\
    .query('FLAG_HP == 1 & count >= 3')\
    .head(20)

df_train\
    .query('`User-ID` == 252829')\
    .sort_values(by = 'Book-Title')
