==> 02-notebooks/03-collaborative-filter/03-comportamento-dados.ipynb <==
# Comportamento dos dados

## 0. Setup

#---- Bibliotecas:

# Análise e manipulação de dados:

import pandas as pd
import numpy as np
from os import chdir

#---- Mudando o diretório raiz para a pasta principal do projeto

chdir('../../')

#---- Carregando as funções

from functions import *

## 1. Extração dos dados

train_data = pd.read_parquet('03-data/02-processed/05-ALS-df-train-data.parquet')

train_data

## 2. Criação da flag de avaliação de livros do Harry Potter

train_data['FLAG_BOOK_HP'] = np.where(train_data['Book-Title'].str.contains('Potter'), 1, 0)

# list_of_hp_books = train_data\
#     .query('FLAG_BOOK_HP == 1')['Book-Title']\
#     .unique()\
#     .tolist()

list_of_hp_books = ['Harry Potter and the Chamber of Secrets (Book 2)',
                    'Harry Potter and the Prisoner of Azkaban (Book 3)',
                    'Harry Potter and the Goblet of Fire (Book 4)',
                    'Harry Potter and the Order of the Phoenix (Book 5)',
                    "Harry Potter and the Sorcerer's Stone (Book 1)"]

list_of_hp_books

pd.set_option('max_colwidth', 800)

for book in list_of_hp_books:

    print(f'>>>>>>>>>>>>> Livro escolhido: {book}')

    # DataFrame temporário com os usuários que avaliaram o livro do Harry Potter

    df_temp = train_data\
        .query(f'`Book-Title` == "{book}"')

    # Lista dos usuários 

    list_users_temp = df_temp['User-ID'].unique().tolist()

    # 

    df_train_data_temp = train_data\
        .query(f'`User-ID`.isin({list_users_temp})')['Book-Title']\
        .value_counts()\
        .reset_index()\
        .query(f'`Book-Title` != "{book}"')\
        .head(20)

    display(df_train_data_temp)

lst_temp = train_data\
    .query('FLAG_BOOK_HP == 1')['Book-Title']\
    .unique()\
    .tolist()

[book for book in lst_temp if 'Book 6' in book]

#---- Não temos um Livro 6 de Harry Potter

# Problema de negócio - O que houve para não ter esse livro?
# Hipóteses:
# - A editora não lançou a partir do livro 6
# - Os livros esgotaram e não entraram para o dataset
# - A nossa amostra pode estar viesada de alguma forma 

train_data\
    .query('`Book-Title`.str.contains("Book 6", na = False)')\
    .query('`Book-Title`.str.contains("Potter")')['Book-Title']\
    .unique()

# Também não temos o livro 6 

train_data\
    .query('`Book-Title`.str.contains("Prince", na = False)')\
    .query('`Book-Title`.str.contains("Potter")')['Book-Title']\
    .unique()
